+++ src/assignment_4_Final/IRandomValueGenerator.java
@@ -2,6 +2,6 @@ package assignment_4_Final;
 
 public interface IRandomValueGenerator {
 
-	public int randNumInRange(int min, int max);
+	public double randNumInRange(double min, double max);
 	public boolean getTrueWithProbability(double probability);
 }
